Add files via upload
The very beginnings of a front-end for a website. The working features on it are: drop-down menu for possible links, an interactive search bar that pops up, and goes away by clicking the "search" button on the navigation bar, and an interactive navigation bar that will eventually lead to other pages on the site.